# server/src/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Property Management API
  description: A sample API

servers:
  - url: http://{base_url}/{env_id}
    variables:
      base_url:
        default: localhost:3000
        description: server
      env_id:
        default: v1
        description: path selector

paths:
  /ping:
    get:
      summary: Ping the API
      operationId: ping
      x-eov-operation-handler: ping
      responses:
        "200":
          description: A greeting for the user.
          content:
            text/plain:
              schema:
                type: string

  /reservations:
    get:
      summary: Get all reservations
      operationId: getAllReservations
      x-eov-operation-handler: reservations
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    
    post:
      summary: Create a new reservation
      operationId: createReservation
      x-eov-operation-handler: reservations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reservation"
      responses:
        "201":
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  /reservations/{id}:
    get:
      summary: Get a reservation by ID
      operationId: getReservationById
      x-eov-operation-handler: reservations
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Reservation found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          description: Reservation not found

    put:
      summary: Update a reservation by ID
      operationId: updateReservationById
      x-eov-operation-handler: reservations
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reservation"
      responses:
        "200":
          description: Reservation updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          description: Reservation not found

    delete:
      summary: Delete a reservation by ID
      operationId: deleteReservationById
      x-eov-operation-handler: reservations
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Reservation deleted successfully
        "404":
          description: Reservation not found

  /properties:
    get:
      summary: Get all properties
      operationId: getAllProperties
      x-eov-operation-handler: properties
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    
    post:
      summary: Create a new property
      operationId: createProperty
      x-eov-operation-handler: properties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Property"
      responses:
        "201":
          description: Property created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"

  /properties/{id}:
    get:
      summary: Get a property by ID
      operationId: getPropertyById
      x-eov-operation-handler: properties
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Property found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Property not found

    put:
      summary: Update a property by ID
      operationId: updatePropertyById
      x-eov-operation-handler: properties
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Property"
      responses:
        "200":
          description: Property updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Property not found

    delete:
      summary: Delete a property by ID
      operationId: deletePropertyById
      x-eov-operation-handler: properties
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Property deleted successfully
        "404":
          description: Property not found

components:
  schemas:
    Reservation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        propertyId:
          type: integer
          format: int64

    Property:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string